Rename `resolver.rs` to `imports.rs`
